Refactor mesh data structure
I rethought what I needed from the data structure and came up with this
as a better design. The new design passes the tests for a first order
mesh (no side or internal nodes) but I haven't implemented the
insertion of the higher order nodes yet, so it fails the test that
requires that functionality.