Number node adjacencies correctly
Tests do not work correctly off of this commit because I don't correctly
update the nodes on shared faces, and this results in an exception
thrown. However, I believe that all of the node adjacencies are updated
correctly and all I need to do is add the logic to update node numbers
on adjacent elements.