A couple of things
I removed the CoordT template argument for BoundaryRepresentation,
as it isn't used at all. I also added a call to `shrink_to_fit` at the
end of assign_face_and_internal_nodes, so that the storage
used by the mesh is the same as the storage used after renumbering
nodes.